+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii 2 File DB extension Change Log
 1.0.1 under development
 -----------------------
 
+- Enh #3: `FileManagerPhp::writeData()` now invalidates script file cache performed by 'OPCache' or 'APC' (klimov-paul)
 - Bug #2: Fixed `ActiveRecord` looses not 'dirty' data on update (fps01)